+++ include/flydsl/Conversion/CMakeLists.txt
@@ -1,8 +1,1 @@
-set(LLVM_TARGET_DEFINITIONS Passes.td)
-mlir_tablegen(Passes.h.inc -gen-pass-decls -name Conversion)
-mlir_tablegen(Passes.capi.h.inc -gen-pass-capi-header --prefix Conversion)
-mlir_tablegen(Passes.capi.cpp.inc -gen-pass-capi-impl --prefix Conversion)
-
-add_mlir_generic_tablegen_target(FlyConversionPassIncGen)
-
-
+add_subdirectory(FlyToROCDL)
