+++ include/flydsl/Conversion/Passes.h
@@ -10,7 +10,7 @@
 namespace mlir {
 
 #define GEN_PASS_REGISTRATION
-#include "flydsl/Conversion/Passes.h.inc"
+#include "flydsl/Conversion/FlyToROCDL/Passes.h.inc"
 
 } // namespace mlir
